requirements.txt
flask==3.0.0
flask-cors==4.0.0
pdfplumber==0.10.3
python-docx==1.1.0
python-dateutil==2.8.2
